Added error handling to business logic [#15]

+++ src/main/java/com/muamerseljubac/service/CategoryService.java
@@ -1,7 +1,7 @@
 package com.muamerseljubac.service;
 
-import com.muamerseljubac.entity.dtos.CategoryAddRequestDTO;
 import com.muamerseljubac.entity.dtos.CategoryDTO;
+import com.muamerseljubac.entity.dtos.request.CategoryAddRequestDTO;
 import com.muamerseljubac.entity.dtos.request.CategoryEditRequestDTO;
 import com.muamerseljubac.entity.dtos.response.CategoryDeleteResponseDTO;
 import com.muamerseljubac.entity.models.Category;
@@ -10,6 +10,7 @@ import com.muamerseljubac.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.data.mapping.PropertyReferenceException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -35,25 +36,41 @@ public class CategoryService {
     }
 
     public CategoryDTO getCategory(UUID id) {
-        return categoryMapper.categoryToCategoryDto(categoryRepository.findById(id).orElse(null));
+        if (categoryRepository.existsById(id)) {
+            return categoryMapper.categoryToCategoryDto(categoryRepository.findById(id).orElse(null));
+        } else {
+            throw new RuntimeException("Category not found!");
+        }
     }
 
     public List<CategoryDTO> getAllCategories(int page, String sort) {
-        return categoryRepository.findAll(PageRequest.of(page, 10, sort != null ? Sort.by(sort) : Sort.by("name")))
-                .stream()
-                .map(categoryMapper::categoryToCategoryDto)
-                .toList();
+        try {
+            return categoryRepository.findAll(PageRequest.of(page, 10, sort != null ? Sort.by(sort) : Sort.by("name")))
+                    .stream()
+                    .map(categoryMapper::categoryToCategoryDto)
+                    .toList();
+        } catch (PropertyReferenceException pre) {
+            throw new RuntimeException("Sort parameter invalid!");
+        }
     }
 
     public CategoryDTO editCategory(CategoryEditRequestDTO requestDTO) {
-        Category category = categoryMapper.categoryEditRequestDtoToCategory(requestDTO);
-        categoryRepository.save(category);
-        return categoryMapper.categoryToCategoryDto(category);
+        if (categoryRepository.existsById(requestDTO.getId())) {
+            Category category = categoryMapper.categoryEditRequestDtoToCategory(requestDTO);
+            categoryRepository.save(category);
+            return categoryMapper.categoryToCategoryDto(category);
+        } else {
+            throw new RuntimeException("Category not found!");
+        }
     }
 
 
     public CategoryDeleteResponseDTO deleteCategory(UUID id) {
-        categoryRepository.deleteById(id);
-        return new CategoryDeleteResponseDTO("Category with ID " + id + " has been deleted successfully!");
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+            return new CategoryDeleteResponseDTO("Category with ID " + id + " has been deleted successfully!");
+        } else {
+            throw new RuntimeException("Category not found!");
+        }
     }
 }
